Update BlockSuperCalc.java
If I am right you forgot to add the check to see if it was not conflicting. In chackNotConficting you returned true if it was conflicting. Forgetting to add the ! means that you are wanting it to conflict. Hope I am right and If I am hope this helps, btw thank you for the code, could really help me out with my multiblocks